Updated bot.js, working and also added flowRate to payload hash.

+++ scripts/bot.js
@@ -3,9 +3,6 @@ const pg = require("pg");
 const format = require("pg-format");
 const connectionString =
     "postgres://superfluid:password@localhost:5432/superfluid";
-const client = new pg.Client({
-    connectionString: connectionString,
-});
 const streamSchedulerAddress = "0x4A679253410272dd5232B3Ff7cF5dbB88f295319";
 
 async function runBot(streamScheduler) {
@@ -18,9 +15,7 @@ async function runBot(streamScheduler) {
         streamScheduler,
         latestBlockNumber,
     );
-
-    // Store all the stream orders back into the database.
-    await storeStreamOrdersIntoDB(pastEvents);
+    console.log("Retrieved past Events, count: ", pastEvents.length);
 
     // Loop through the events and check differentiate the diff types of stream orders.
     await processStreamOrders(streamScheduler, pastEvents);
@@ -35,6 +30,7 @@ const parseEventDataArgs = eventData => {
     streamOrderData.sender = eventData.sender;
     streamOrderData.superToken = eventData.superToken;
     streamOrderData.flowRate = eventData.flowRate;
+    streamOrderData.startTime = eventData.startTime;
     streamOrderData.endTime = eventData.endTime;
     streamOrderData.userData = eventData.userData;
     return streamOrderData;
@@ -56,31 +52,29 @@ async function processStreamOrders(streamScheduler, events) {
     let streamOrdersToDelete = [];
     let streamOrdersToUpdateOrCreate = [];
     for (let i = 0; i < events.length; i++) {
+        let streamOrderData = parseEventDataArgs(events[i].args);
         if (events[i].endTime != 0 && events[i].endTime < timeNowInSecs) {
             console.log("Detected close stream order");
             await streamScheduler.executeDeleteStream(
-                events[i].receiver,
-                events[i].superToken,
-                events[i].startTime,
-                events[i].flowRate,
-                events[i].endTime,
-                events[i].userData,
+                streamOrderData.receiver,
+                streamOrderData.superToken,
+                streamOrderData.startTime,
+                streamOrderData.flowRate,
+                streamOrderData.endTime,
+                streamOrderData.userData,
             );
-            streamOrdersToDelete.push({
-                receiver: events[i].event_receiver,
-                sender: events[i].event_sender,
-            });
+            streamOrdersToDelete.push(events[i]);
         } else {
             streamOrdersToUpdateOrCreate.push({
-                name: events[i].event_name,
-                receiver: events[i].event_receiver,
-                sender: events[i].event_sender,
-                blockNumber: parseInt(events[i].event_block_number),
-                flowRate: parseInt(events[i].event_flow_rate),
-                endTime: parseInt(events[i].event_end_time),
-                startTime: parseInt(events[i].event_start_time),
-                superToken: events[i].event_super_token,
-                userData: events[i].event_user_data,
+                name: events[i].event,
+                receiver: streamOrderData.receiver,
+                sender: streamOrderData.sender,
+                blockNumber: parseInt(events[i].blockNumber),
+                flowRate: parseInt(streamOrderData.flowRate),
+                endTime: parseInt(streamOrderData.endTime),
+                startTime: parseInt(streamOrderData.startTime),
+                superToken: streamOrderData.superToken,
+                userData: streamOrderData.userData,
             });
         }
     }
@@ -91,6 +85,10 @@ async function processStreamOrders(streamScheduler, events) {
             streamOrder,
         );
         if (shouldCreateStreamOrder) {
+            console.log(
+                "Did not find sender/receiver pair, creating stream order with stream order: ",
+                streamOrder,
+            );
             await streamScheduler.executeCreateStream(
                 streamOrder.receiver,
                 streamOrder.superToken,
@@ -100,6 +98,10 @@ async function processStreamOrders(streamScheduler, events) {
                 streamOrder.userData,
             );
         } else {
+            console.log(
+                "Found sender/receiver pair, updating stream order with stream order: ",
+                streamOrder,
+            );
             await streamScheduler.executeUpdateStream(
                 streamOrder.receiver,
                 streamOrder.superToken,
@@ -114,30 +116,48 @@ async function processStreamOrders(streamScheduler, events) {
     }
 
     // Delete stream orders that have been closed from the database.
+    // Store all the stream orders back into the database.
+    await storeStreamOrdersIntoDB(
+        // Don't store the stream orders that will be deleted.
+        events.filter(o1 => {
+            return !streamOrdersToDelete.some(o2 => {
+                return o1.blockNumber === o2.blockNumber;
+            });
+        }),
+    );
     streamOrdersToDelete.concat(await getExpiredStreamOrdersFromDB());
-    await deleteStreamOrdersFromDB(streamOrdersToDelete);
+    if (streamOrdersToDelete.length > 0) {
+        await deleteStreamOrdersFromDB(streamOrdersToDelete);
+    }
 }
 
 async function deleteStreamOrdersFromDB(streamOrdersToDelete) {
+    const client = new pg.Client({
+        connectionString: connectionString,
+    });
+    const queryArray = Array(streamOrdersToDelete.length).fill("(?)");
     const sql = format(
         `DELETE FROM stream_orders 
         WHERE (event_receiver, event_sender)
-        IN %L`,
+        IN (${queryArray.join(",")})`,
         streamOrdersToDelete,
     );
+    console.log("DELETE SQL: ", sql);
     try {
         await client.connect();
         const result = await client.query(sql);
         console.log("Deleted stream orders from database");
     } catch (e) {
-        console.log("Error inserting stream orders into database, " + e);
+        console.log("Error deleting stream orders into database, " + e);
     } finally {
         await client.end();
     }
-    await client.end();
 }
 
 async function getExpiredStreamOrdersFromDB() {
+    const client = new pg.Client({
+        connectionString: connectionString,
+    });
     const timeNowInSecs = Math.floor(Date.now() / 1000);
     const sql = `SELECT event_receiver, event_sender
     FROM stream_orders WHERE event_end_time < $1`;
@@ -146,23 +166,27 @@ async function getExpiredStreamOrdersFromDB() {
     try {
         await client.connect();
         const result = await client.query(sql, values);
+        return parseToDeleteStreamOrder(result.rows);
     } catch (e) {
-        console.log("Error inserting stream orders into database, " + e);
+        console.log("Error querying for expired streams in database, " + e);
+        return [];
     } finally {
         await client.end();
     }
-    return parseToDeleteStreamOrder(result.rows);
 }
 
 async function checkIfStreamOrderExistsInDB(streamOrder) {
+    const client = new pg.Client({
+        connectionString: connectionString,
+    });
     try {
         await client.connect();
         const sql = `select count(*) from stream_orders 
-        where stream_order.receiver = event_receiver
-        and stream_order.sender = event_sender`;
+        where stream_orders.event_receiver = '${streamOrder.receiver}'
+        and stream_orders.event_sender = '${streamOrder.sender}'`;
         const result = await client.query(sql);
         await client.end();
-        return result == 0 ? true : false;
+        return result.rows[0].count == 0 ? true : false;
     } catch (e) {
         console.log(
             "Error checking if stream order exists in database, error: ",
@@ -174,6 +198,9 @@ async function checkIfStreamOrderExistsInDB(streamOrder) {
 }
 
 async function updateStreamOrderInDB(streamOrder) {
+    const client = new pg.Client({
+        connectionString: connectionString,
+    });
     const sql = `UPDATE stream_orders SET
             event_flow_rate = ?,
             WHERE event_sender = ? AND event_receiver = ?`;
@@ -194,7 +221,10 @@ async function updateStreamOrderInDB(streamOrder) {
 }
 
 async function getLatestBlockNumberFromDB() {
-    const sql = `SELECT max(event_block_number) as latest_block_number FROM stream_orders`;
+    const client = new pg.Client({
+        connectionString: connectionString,
+    });
+    const sql = `SELECT max(event_block_number) as event_block_number FROM stream_orders`;
     try {
         await client.connect();
         const result = await client.query(sql);
@@ -204,14 +234,19 @@ async function getLatestBlockNumberFromDB() {
             );
             return 0;
         }
-        console.log("Found data in the database, returning latest block.");
-        return parseInt(result.rows[0].event_block_number);
+        console.log(
+            "Found data in the database, returning latest block, ",
+            result.rows[0],
+        );
+        return parseInt(result.rows[0].event_block_number) + 1;
     } catch (e) {
         console.log(
             "Error getting latest block number from database, error: ",
             e,
         );
+        return 0;
     } finally {
+        console.log("ending");
         await client.end();
     }
 }
@@ -239,26 +274,34 @@ async function storeStreamOrdersIntoDB(events) {
             streamOrderData.receiver,
             streamOrderData.sender,
             streamOrderData.superToken,
-            streamOrderData.flowRate,
-            streamOrderData.endTime,
+            parseInt(streamOrderData.flowRate.toString()),
+            parseInt(streamOrderData.startTime.toString()),
+            parseInt(streamOrderData.endTime.toString()),
             streamOrderData.userData,
         ];
         streamOrderList.push(blob);
     }
     // Insert the stream orders into the database.
-    await insertStreamOrdersIntoDB(streamOrderList);
+    console.log("Stream order list: ", streamOrderList);
+    if (streamOrderList.length > 0) {
+        await insertStreamOrdersIntoDB(streamOrderList);
+    }
 }
 
 async function insertStreamOrdersIntoDB(streamOrderList) {
     // Insert 2d array of stream orders into the database.
+    const client = new pg.Client({
+        connectionString: connectionString,
+    });
     const sql = format(
-        `INSERT INTO stream_order (
-            event_name,
+        `INSERT INTO stream_orders (
+        event_name,
             event_block_number,
             event_receiver,
             event_sender,
             event_super_token,
             event_flow_rate,
+            event_start_time,
             event_end_time,
             event_user_data
             ) VALUES %L`,
@@ -267,13 +310,15 @@ async function insertStreamOrdersIntoDB(streamOrderList) {
     try {
         await client.connect();
         const result = await client.query(sql);
+        console.log(
+            "Inserted stream orders into DB, row count: ",
+            result.rowCount,
+        );
     } catch (e) {
         console.log("Error inserting stream orders into database, " + e);
     } finally {
         await client.end();
     }
-
-    console.log("Inserted stream orders into DB: ", result);
 }
 
 async function main() {
